+++ src/components/SearchResultsList/SearchResultsList.js
@@ -1,10 +1,10 @@
-import { useMemo } from 'react';
+import {
+  useMemo,
+  useContext,
+} from 'react';
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
-import {
-  useLocation,
-  useRouteMatch,
-} from 'react-router';
+import { useRouteMatch } from 'react-router';
 
 import {
   MultiColumnList,
@@ -12,14 +12,17 @@ import {
 } from '@folio/stripes/components';
 import { SearchAndSortNoResultsMessage } from '@folio/stripes/smart-components';
 
+import { SelectedAuthorityRecordContext } from '../../context';
+
 import { AuthorityShape } from '../../constants/shapes';
-import {
-  searchResultListColumns,
-} from '../../constants';
+import { searchResultListColumns } from '../../constants';
+
+import css from './SearchResultsList.css';
 
 const propTypes = {
   authorities: PropTypes.arrayOf(AuthorityShape).isRequired,
   hasFilters: PropTypes.bool.isRequired,
+  hidePageIndices: PropTypes.bool,
   isFilterPaneVisible: PropTypes.bool.isRequired,
   loaded: PropTypes.bool.isRequired,
   loading: PropTypes.bool.isRequired,
@@ -51,10 +54,12 @@ const SearchResultsList = ({
   query,
   toggleFilterPane,
   hasFilters,
+  hidePageIndices,
 }) => {
   const intl = useIntl();
-  const location = useLocation();
   const match = useRouteMatch();
+
+  const [selectedAuthorityRecordContext, setSelectedAuthorityRecordContext] = useContext(SelectedAuthorityRecordContext);
 
   const columnMapping = {
     [searchResultListColumns.AUTH_REF_TYPE]: intl.formatMessage({ id: 'ui-marc-authorities.search-results-list.authRefType' }),
@@ -76,18 +81,23 @@ const SearchResultsList = ({
     },
     headingRef: (authority) => (
       <TextLink
-        to={`${match.path}/authorities/${authority.id}${location.search}`}
+        to={`${match.path}/authorities/${authority.id}`}
+        className={authority.isAnchor ? css.anchorLink : null}
       >
         {authority.headingRef}
       </TextLink>
     ),
   };
 
+  const onRowClick = (e, row) => {
+    setSelectedAuthorityRecordContext(row);
+  };
+
   const source = useMemo(
     () => ({
       loaded: () => (hasFilters || query) && loaded,
       pending: () => loading,
-      failure: () => { },
+      failure: () => null,
     }),
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [loading, hasFilters],
@@ -102,6 +112,8 @@ const SearchResultsList = ({
       id="authority-result-list"
       onNeedMoreData={onNeedMoreData}
       visibleColumns={visibleColumns}
+      selectedRow={selectedAuthorityRecordContext}
+      onRowClick={onRowClick}
       totalCount={totalResults}
       pagingType="prev-next"
       pageAmount={pageSize}
@@ -110,6 +122,7 @@ const SearchResultsList = ({
       sortOrder={sortOrder}
       onHeaderClick={onHeaderClick}
       autosize
+      hidePageIndices={hidePageIndices}
       isEmptyMessage={
         source ? (
           <div data-test-agreements-no-results-message>